+++ CHANGELOG.md
@@ -1,13 +1,23 @@
 # Change Log
 
+### v0.0.2.0
+Setup the threading system for Panther `Context`. Allows for both single-threading and multi-threading
+##### pthr
+- Setup basic runtime config settings
+##### Panther
+- Added a number of functions to `Context` related to the threading and tasks
+- Added `Context::loadFiles()`
+##### Misc
+- Added the testing directory
+
 ### v0.0.1.0
 Setting up basic frameworks
-#### Panther
+##### Panther
 - Added `Context`
 - Added `Source`
 - Added `SourceManager`
 - Added diagnostics
-#### PCIT_core
+##### PCIT_core
 - Added `DiagnosticImpl`
 - Added `Printer`
 - Added `UniqueID` and `UniqueComparableID`